Moved Authorization folder from Library to ApiService project. ControllerWithClientAuthorization was also moved to ApiService project
