+++ src/OnlineShop.ApiService/OnlineShop.ApiService.csproj
@@ -7,6 +7,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="IdentityModel" Version="6.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
     <ProjectReference Include="..\OnlineShop.Library\OnlineShop.Library.csproj" />
   </ItemGroup>
